Pretty much initial commit
And so I'm back with the portion of some weird and pathetic code. This time around, that's gonna be an application that will render an image or set of images to their "grayscale byte equivalent" and then do stuff with it. At least for now, "stuff" is intended to be an ability to translate bytes into ASCII characters and display it on the console. The goal is to achieve 24FPS at a decent resolution (and 60FPS at medium for sure). But I hope that I will leave a possibility to do other ways of exporting rendered grayscale elsewhere. Yikes!